feat: Moar refactor
