+++ src/auth_browser/mod.rs
@@ -1,1 +1,2 @@
 pub mod auth_browser;
+pub mod auth_page;
